use raw scalar for open
Perl 5.22 introduced a bug when opening file handles into a scalar ref.
An extra reference to the handle is created preventing the handle from
being closed when the variable goes out of scope.  For scripts that open
a large number of files, this can exhaust the OS limit on open file
descriptors.  Work around the issue by opening into a plain scalar and
assigning to the destination following the open.

Fixes #3